// backend/src/common/enums/transfer-status.enum.ts
export enum TransferStatus {
  PENDIENTE = 'PENDIENTE',
  ASIGNADA = 'ASIGNADA',
  EN_PREPARACION = 'EN_PREPARACION',
  LISTA_DESPACHO = 'LISTA_DESPACHO',
  EN_TRANSITO = 'EN_TRANSITO',
  LLEGADA_DESTINO = 'LLEGADA_DESTINO',
  COMPLETADA = 'COMPLETADA',
  COMPLETADA_CON_DISCREPANCIA = 'COMPLETADA_CON_DISCREPANCIA',
  CANCELADA = 'CANCELADA',
}